feat: ItemManager 생성
